ugly but functional
