+++ client/src/utils/queries.js
@@ -23,8 +23,8 @@ export const QUERY_USER = gql`
 `;
 
 export const QUERY_CHECKOUT = gql`
-  query getCheckout($products: [ID]!) {
-    checkout(products: $products) {
+  {
+    checkout {
       session
     }
   }
